V1.6.2
Major Changes to js. Converted from Jq to JS for greater control, independence and more room for creativity with animations etc.

Also fixed one or two issues/bugs and cleaned up the file structure.

Next up will be animations and UX

### Bug fixes ###

Cardsection now dissapears when new tab is selected.
header is no longer visible on the entry prompt on mobile.
Sizing is now grabbed and adjusted for mobile when selecting a new tab. Previously the height was static which caused spacing issues for all tabs on mobile.